Smarter AMD export
Currently the AMD transform only attempts to replace the trailing
node in the definition iif it's a `ReturnStatement`. If the `return`
statement appears else where in the definition no `export` is
produced.

With this patch the candidate `return` node is the last `return`
node that appears in the definition's body.